+++ src/modules/eliteboard/types/Port_fbt.h
@@ -14,7 +14,7 @@
 #define _PORT_H_
 
 #include <array>
-#include "PortAdapter.h"
+#include "PortAdapter_atp.h"
 #include "forte_array_at.h"
 #include "forte_wstring.h"
 #include "funcbloc.h"
